requirements_dev.txt
catboost>=1.1.1
deap>=1.3.3
joblib>=1.2.0
keras==2.12.0
pandas>=1.5.3
python-dateutil>=2.8.1
pytz>=2022.7.1
scikit-learn>=1.2.1
scipy>=1.10.0
seaborn==0.12.2
six>=1.15.0
tensorflow>=2.12.0
xgboost>=1.7.3
plotly>=5.15.0
pytest-cov
kaleido
sphinx-gallery==0.14.0
sphinx_book_theme
sphinx_mdinclude
sphinx-autoapi
sphinx-favicon
mlflow
sphinxcontrib-mermaid